#56: Support async/await in Rust -> Java direction.
Implemented the invoke_async function that returns a Future, which is completed via the Receiver of a oneshot channel. In Java side, the methods that are allowed to be invoked by invoke_async, are the ones that return a Java Future. When the Java Future completes, the Java side of j4rs invokes native Rust code that completes the Rust pending Future with either success or failure, using the Sender of the oneshot channel that was created when the invoke_async was called.

Please note that it is best that the Java methods that are invoked by invoke_async function return a CompletableFuture, as this improves performance.
j4rs handles simple Java Futures with polling, using an internal one-threaded ScheduledExecutorService. This has apparent performance issues.