Prefer overloads with arguments with specific types over generics during method lookup
When a class has a method that has multiple overloads, prefer overloads where
arguments are of specific types over overloads that accept generic types.

To give a more concrete example, if I have the following overloads:

```
public void myMethod(Integer value) { ... } // overload 1
public<T> void myMethod(T value) { ... }    // overload 2
```

this change ensures that we always try to match parameter types against the
overload that accepts `Integer` first. Since `T` matches anything, if we
were to check the overload 2 first, we would always match it, even if
the argument we are trying to match is of type `Integer` - this would
result in us calling the wrong overload!

This order of matching is more consistent with what Java does when
resolving calls to overloaded methods.